posTimer in AdminPlayer neu: Timer in AVPlayer wurde nicht installliert (ev anderer Thread)
setZeilenAufnahme neu: setzt Aufnahme in AdminPlayer 'Aufnahmen nach Namen'

setAdminPlayer: inProjekt: Pfad mit m4a/txt

startAVPlay: posTimer in AdminPlayer installiert, add runloop war notwendig, fkt. setAbspielanzeigeanPos neu, ohne Notification

+++ RecPlayAdminPlayer/rAdminDS.m
@@ -340,7 +340,7 @@ return index;
 			while(eineAufnahme=[AufnahmenEnumerator nextObject])//Aufnahmen für Menu
          {
             [cell addItemWithTitle:eineAufnahme];
-            int menuIndex=[cell indexOfItemWithTitle:eineAufnahme];
+            double menuIndex=[cell indexOfItemWithTitle:eineAufnahme];
             //NSLog(@"eineAufnahme: %@ index: %d  menuIndex: %d",eineAufnahme,index,menuIndex);
             
             if ([[MarkArray objectAtIndex:row]count])
